requirements.txt
deepdiff==1.0.0
Django==1.9.5
django-audit-log==0.7.0
django-countries==3.4.1
django-daterange-filter==1.2.0
django-languages==0.1
django-localflavor==1.2
django-simple-history==1.8.0
django-tz-detect==0.2.6
django-widget-tweaks==1.4.1
fdfgen==0.12.1
PyMySQL==0.7.2
pypiwin32==219
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
xlwt==1.0.0
django-import-export==0.4.5